Create util.h
Custom util.h with an additional IFDEF included to check the GLIBC version being used, and in cases where necessary, import the syscall.h header in order to be able to use the gettid function.

Without this ifdef, snort 2.9.14.1 fails to compile on Ubuntu 24.04. Unsure what older versions of snort this affects. It doesn't not affect 2.9.15.1 and newer, nor does it appear to affect version 2.9.9.0